# Faster IndexOfAny for IgnoreCase Ascii letters