--- a/src/libraries/System.Private.CoreLib/src/System/SearchValues/Any1CharPackedIgnoreCaseSearchValues.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/SearchValues/Any1CharPackedIgnoreCaseSearchValues.cs
@@ -14,7 +14,6 @@ namespace System.Buffers
 
         public Any1CharPackedIgnoreCaseSearchValues(char value)
         {
-            Debug.Assert(value != 0);
             Debug.Assert((value | 0x20) == value);
 
             _lowerCase = value;