--- a/src/libraries/System.Private.CoreLib/src/System/SearchValues/Any2SearchValues.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/SearchValues/Any2SearchValues.cs
@@ -31,8 +31,7 @@ namespace System.Buffers
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         internal override unsafe bool ContainsCore(T value) =>
-            *(TImpl*)&value == _e0 ||
-            *(TImpl*)&value == _e1;
+            *(TImpl*)&value == _e0 || *(TImpl*)&value == _e1;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         internal override int IndexOfAny(ReadOnlySpan<T> span) =>