… src/libraries/System.Memory/tests/Span/SearchValues.cs
… src/libraries/System.Private.CoreLib/src/System/SearchValues/Any2ByteSearchValues.cs
… src/libraries/System.Private.CoreLib/src/System/SearchValues/Any2CharSearchValues.cs
… src/libraries/System.Private.CoreLib/src/System/SearchValues/Any3ByteSearchValues.cs
… src/libraries/System.Private.CoreLib/src/System/SearchValues/Any3CharSearchValues.cs
… src/libraries/System.Private.CoreLib/src/System/SearchValues/SearchValues.cs
… src/libraries/System.Private.CoreLib/src/System/SearchValues/SingleByteSearchValues.cs
… src/libraries/System.Private.CoreLib/src/System/SearchValues/SingleCharSearchValues.cs
… src/libraries/System.Private.CoreLib/src/System/SpanHelpers.Packed.cs
… src/libraries/System.Private.CoreLib/src/System/SpanHelpers.T.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexCompiler.cs