The idea is that if we're searching for both cases of an Ascii letter, we can do it by masking off the bit where they differ and only search for one. E.g. `(input | 0x20) is 'a' or 'b'` instead of `input is 'a' or 'A' or 'b' or 'B'`.

This PR adds a variant of packed `IndexOfAny(char)` and `IndexOfAny(char, char)` that does the `| 0x20` input transformation and teaches `SearchValues` to use it.
I've also updated the `RegexCompiler` to use `SearchValues` for sets of 4/5 values like the source generator does, so that it can make use of this change.

```c#
public class SearchValuesPackedIgnoreCase
{
    private static readonly SearchValues<char> s_asciiTwoLetters = SearchValues.Create("Aa");
    private static readonly SearchValues<char> s_asciiFourLetters = SearchValues.Create("AaBb");
    private static readonly string s_text = new string('\n', 20_000);

    [Benchmark] public int IndexOfOneIgnoreCase() => s_text.AsSpan().IndexOfAny(s_asciiTwoLetters);
    [Benchmark] public int IndexOfTwoIgnoreCase() => s_text.AsSpan().IndexOfAny(s_asciiFourLetters);
}
```

| Method                 | Toolchain | Mean     | Error   | Ratio |
|----------------------- |-----------|---------:|--------:|------:|
| IndexOfOneIgnoreCase   |  main     | 583.5 ns | 3.56 ns |  1.00 |
| IndexOfOneIgnoreCase   |  pr       | 499.5 ns | 1.71 ns |  0.86 |
|                        |           |          |         |       |
| IndexOfTwoIgnoreCase   |  main     | 742.5 ns | 4.02 ns |  1.00 |
| IndexOfTwoIgnoreCase   |  pr       | 642.2 ns | 5.54 ns |  0.87 |
																					  
| Method | Toolchain | Pattern             | Options  | Mean      | Error    | Ratio |
|------- |-----------|-------------------- |--------- |----------:|---------:|------:|
| Count  |  main     | (?i)Sherlock\|Holmes | Compiled | 676.61 us | 1.858 us |  1.00 |
| Count  |  pr       | (?i)Sherlock\|Holmes | Compiled | 623.75 us | 1.762 us |  0.92 |
|        |           |                     |          |           |          |       |
| Count  |  main     | [Zz]                | Compiled |  32.16 us | 0.126 us |  1.00 |
| Count  |  pr       | [Zz]                | Compiled |  30.39 us | 0.112 us |  0.95 |
|        |           |                     |          |           |          |       |
| Count  |  main     | [ZzQq]              | Compiled |  63.99 us | 0.289 us |  1.00 |
| Count  |  pr       | [ZzQq]              | Compiled |  46.14 us | 0.131 us |  0.72 |

The bigger difference for `[ZzQq]` is switching from a regular 4-value `IndexOfAny` to a 2-value packed one.

I tested the same approach for `IndexOfAny(char, char, char)` and `IndexOfAnyInRange`, but the throughput there is pretty much on par with or slower than `IndexOfAnyAsciiSearcher`, so I avoided adding more code for it.